minimized changes to python-support especially not using venv
