+++ CMakeLists.txt
@@ -591,7 +591,6 @@ set(CORE_SOURCES
   src/core/Expression.cc
   src/core/builtin_functions.cc
   src/core/function.cc
-  src/core/FunctionType.cc
   src/core/ImportNode.cc
   src/core/LinearExtrudeNode.cc
   src/core/LocalScope.cc
@@ -612,7 +611,6 @@ set(CORE_SOURCES
   src/core/SurfaceNode.cc
   src/core/TextNode.cc
   src/core/TransformNode.cc
-  src/core/UndefType.cc
   src/core/Value.cc
   src/core/Assignment.cc
   src/core/AST.cc
