+++ doc/testing.txt
@@ -5,7 +5,7 @@ Running regression tests:
 
 Install the prerequisite helper programs on your system:
 
-    cmake, python3, python3-venv, python3-pip
+    cmake, python3, ImageMagick 6.5.9.3 or newer
 
 There are binary installer packages of these tools available for Mac,
 Win, Linux, BSD, and other systems.
@@ -124,21 +124,20 @@ A pretty-printed index.html is in a subdir of tests/build/Testing/Temporary
 Expected results are found in tests/regression/*
 Actual results are found in tests/build/testname-output/*
 
-3. Alternatives to the image_compare.py image comparison script:
+3. Image-based tests takes a long time, they fail, and the log says 'return -11'
 
-If cmake is given the option -DUSE_IMAGE_COMPARE_PY=OFF then ImageMagick
-comparison and fallback to diffpng are available.  Note that ImageMagick
-tests are less sensitive because they are pixel-based with a large threshold
-while image_compare.py checks for any 3x3 blocks (with overlap) that have
-non-zero differences of the same sign.
+Imagemagick may have crashed while comparing the expected images to the 
+test-run generated (actual) images. You can try using the alternate 
+ImageMagick comparison method by by erasing CMakeCache, and re-running 
+cmake with -DCOMPARATOR=ncc. This will enable the Normalized Cross 
+Comparison method which is less accurate but won't usually crash.
 
-With -DUSE_IMAGE_COMPARE_PY=OFF additional options are available:
- : -DCOMPARATOR=ncc Normalized Cross Comparison which is less accurate but
-   more runtime stable on some ImageMagick versions.
- : -DCOMPARATOR=old Lowered reliability test that works on older
-   ImageMagick versions.  Use this with "morphology not found" in the log.
+4. Testing images fails with 'morphology not found" for ImageMagick in the log
 
-4. Locale errors
+Your version of imagemagick is old. Upgrade, or pass -DCOMPARATOR=old to 
+cmake. The comparison will be of lowered reliability.
+
+5. Locale errors
 
 "terminate called after throwing an instance of 'std::runtime_error'
   what():  locale::facet::_S_create_c_locale name not valid" 
@@ -147,7 +146,7 @@ Is a boost/libstdc++ bug. Fix like so before running:
 
   $ export LC_MESSAGES=
 
-5. I want to build without OpenGL
+6. I want to build without OpenGL
 
  There is an unsupported way to do this, by defining NULLGL to Cmake:
 
@@ -158,13 +157,13 @@ Is a boost/libstdc++ bug. Fix like so before running:
  useful for debugging and outputting 3d-formats like STL on systems without GL.
  This option may break in the future and require tweaking to get working again.
 
-6. Proprietary GL driver issues
+7. Proprietary GL driver issues
 
 There are sporadic reports of problems running on remote machines with
 proprietary GL drivers. Try doing a web search for your exact error
 message to see solutions and workarounds that others have found.
 
-7. Windows + MSVC: 
+8. Windows + MSVC: 
 
 The MSVC build was last tested circa 2012. The last time it worked,
 these were the necessary commands to run.
@@ -177,7 +176,7 @@ these were the necessary commands to run.
 > cmake .
 > nmake -f Makefile
 
-8. Other issues
+9. Other issues
 
 The OpenSCAD User Manual Wiki has a section on buildling. Please check 
 there for possible updates and workarounds:
