+++ scripts/msys2-install-dependencies.sh
@@ -43,8 +43,6 @@ pacboy --noconfirm --sync --needed \
     ghostscript:p \
     imagemagick:p \
     qt5-svg:p \
-    qt5-multimedia:p \
-    python:p \
-    python-pip:p
+    qt5-multimedia:p
 
 date "+### %Y-%m-%d %T msys2-install-dependencies finished"
