--- src/core/UndefType.cc
@@ -1,14 +1,0 @@
-#include "Value.h"
-
-Value UndefType::operator<(const UndefType& /*other*/) const {
-  return Value::undef("operation undefined (undefined < undefined)");
-}
-Value UndefType::operator>(const UndefType& /*other*/) const {
-  return Value::undef("operation undefined (undefined > undefined)");
-}
-Value UndefType::operator<=(const UndefType& /*other*/) const {
-  return Value::undef("operation undefined (undefined <= undefined)");
-}
-Value UndefType::operator>=(const UndefType& /*other*/) const {
-  return Value::undef("operation undefined (undefined >= undefined)");
-}
