+++ tests/CMakeLists.txt
@@ -13,8 +13,6 @@ cmake_policy(SET CMP0070 NEW) # Allow relative paths for file(GENERATE ...)
 
 enable_testing()
 
-option(USE_IMAGE_COMPARE_PY "Use built-in image_compare.py" ON)
-
 # Use variables for common paths, to reduce line lengths.
 
 # Very commonly used cmake-provided paths
@@ -54,103 +52,54 @@ find_package(PythonInterp 3.4 REQUIRED)
 
 # Image comparison - expected test image vs actual generated image
 
-if(USE_IMAGE_COMPARE_PY)
-  set(VENV_DIR "${CCBD}/venv")
-  message(STATUS "Preparing image_compare.py for test suite image comparison")
-  if(WIN32)
-    set(IMAGE_COMPARE_EXE "${VENV_DIR}/Scripts/python.exe")
-  else()
-    set(IMAGE_COMPARE_EXE "${VENV_DIR}/bin/python")
-  endif()
-  if(EXISTS "${IMAGE_COMPARE_EXE}")
-    message(STATUS "venv found, testing libraries")
-    execute_process(
-      COMMAND "${IMAGE_COMPARE_EXE}" "${CCSD}/image_compare.py" "--status"
-      WORKING_DIRECTORY "${CCSD}" ERROR_QUIET RESULT_VARIABLE ret)
-    if(ret AND NOT ret EQUAL 0)
-      message(STATUS "venv libraries incomplete")
-      set(BUILD_VENV TRUE)
-    else()
-      message(STATUS "venv libraries complete")
-      set(BUILD_VENV FALSE)
-    endif()
-  else()
-    message(STATUS "venv not found")
-    set(BUILD_VENV TRUE)
-  endif()
-  if(BUILD_VENV)
-    message(STATUS "Setting up testing venv for image comparison")
-    find_package(Python3 COMPONENTS Interpreter)
-    execute_process(
-      COMMAND "${Python3_EXECUTABLE}" "-m" "venv" "venv" "--without-pip"
-      WORKING_DIRECTORY "${CCBD}")
-    execute_process(
-      COMMAND "${IMAGE_COMPARE_EXE}" "-m" "ensurepip"
-      WORKING_DIRECTORY "${CCBD}")
-    execute_process(
-      COMMAND "${IMAGE_COMPARE_EXE}" "-m" "pip" "install" "numpy" "Pillow"
-      WORKING_DIRECTORY "${CCBD}")
-  endif()
-  set(COMPARATOR "--comparator=image_compare")
-  if(BUILD_VENV)
-    execute_process(
-      COMMAND "${IMAGE_COMPARE_EXE}" "${CCSD}/image_compare.py" "--status"
-      WORKING_DIRECTORY "${CCSD}" RESULT_VARIABLE ret)
-    if(ret AND NOT ret EQUAL 0)
-      message(WARNING "Failed to setup the test suite venv for ${IMAGE_COMPARE_EXE}  See doc/testing.txt for dependency information.")
-    else()
-      message(STATUS "venv setup for ${IMAGE_COMPARE_EXE}")
-    endif()
-  else()
-    message(STATUS "venv already setup for ${IMAGE_COMPARE_EXE}")
-  endif()
+# Imagemagick
+find_package(ImageMagick COMPONENTS convert)
+if (ImageMagick_convert_FOUND)
+  message(STATUS "ImageMagick convert executable found: " ${ImageMagick_convert_EXECUTABLE})
+  set(IMAGE_COMPARE_EXE ${ImageMagick_convert_EXECUTABLE})
 else()
-  # Imagemagick
-  find_package(ImageMagick COMPONENTS convert)
-  if(ImageMagick_convert_FOUND)
-    message(STATUS "ImageMagick convert executable found: " ${ImageMagick_convert_EXECUTABLE})
-    set(IMAGE_COMPARE_EXE ${ImageMagick_convert_EXECUTABLE})
-    if ("${ImageMagick_VERSION_STRING}" VERSION_LESS "6.5.9.4")
-      message(STATUS "ImageMagick version less than 6.5.9.4, cannot use -morphology comparison")
-      message(STATUS "ImageMagick Using older image comparison method")
-      set(COMPARATOR "--comparator=old")
-    endif()
+  message(STATUS "Couldn't find imagemagick 'convert' program")
+  set(DIFFPNG 1)
+endif()
 
-    execute_process(COMMAND ${IMAGE_COMPARE_EXE} --version OUTPUT_VARIABLE IM_OUT)
-    if (${IM_OUT} MATCHES "OpenMP")
-      # http://www.daniloaz.com/en/617/systems/high-cpu-load-when-converting-images-with-imagemagick
-      message(STATUS "ImageMagick: OpenMP bug workaround - setting MAGICK_THREAD_LIMIT=1")
-      list(APPEND CTEST_ENVIRONMENT "MAGICK_THREAD_LIMIT=1")
-    endif()
+if (ImageMagick_convert_FOUND)
+  if ("${ImageMagick_VERSION_STRING}" VERSION_LESS "6.5.9.4")
+    message(STATUS "ImageMagick version less than 6.5.9.4, cannot use -morphology comparison")
+    message(STATUS "ImageMagick Using older image comparison method")
+    set(COMPARATOR "--comparator=old")
+  endif()
 
-    message(STATUS "Comparing magicktest1.png with magicktest2.png")
-    set(IM_TEST_FILES "${CCSD}/magicktest1.png" "${CCSD}/magicktest2.png")
-    set(COMPARE_ARGS ${IMAGE_COMPARE_EXE} ${IM_TEST_FILES} -alpha On -compose difference -composite -threshold 8% -morphology Erode Square -format %[fx:w*h*mean] info:)
-    # compare arguments taken from test_cmdline_tool.py
-    message(STATUS "Running ImageMagick compare: ${COMPARE_ARGS}")
-    execute_process(COMMAND ${COMPARE_ARGS} RESULT_VARIABLE IM_RESULT OUTPUT_VARIABLE IM_OUT)
-    if (IM_RESULT)
-      message(STATUS "ImageMagick failed to run")
-      message(STATUS "Using alternative image comparison")
-      set(DIFFPNG 1)
-    elif (${IM_OUT} STREQUAL "0")
-       message(STATUS "magicktest1.png and magicktest2.png were incorrectly detected as identical")
-       message(STATUS "Using alternative image comparison")
-       set(DIFFPNG 1)
-    else()
-      message(STATUS "ImageMagick OK: Detected pixel difference of ${IM_OUT}")
-    endif()
-  else()
-    message(STATUS "Couldn't find imagemagick 'convert' program")
+  execute_process(COMMAND ${IMAGE_COMPARE_EXE} --version OUTPUT_VARIABLE IM_OUT)
+  if (${IM_OUT} MATCHES "OpenMP")
+    # http://www.daniloaz.com/en/617/systems/high-cpu-load-when-converting-images-with-imagemagick
+    message(STATUS "ImageMagick: OpenMP bug workaround - setting MAGICK_THREAD_LIMIT=1")
+    list(APPEND CTEST_ENVIRONMENT "MAGICK_THREAD_LIMIT=1")
+  endif()
+
+  message(STATUS "Comparing magicktest1.png with magicktest2.png")
+  set(IM_TEST_FILES "${CCSD}/magicktest1.png" "${CCSD}/magicktest2.png")
+  set(COMPARE_ARGS ${IMAGE_COMPARE_EXE} ${IM_TEST_FILES} -alpha On -compose difference -composite -threshold 8% -morphology Erode Square -format %[fx:w*h*mean] info:)
+  # compare arguments taken from test_cmdline_tool.py
+  execute_process(COMMAND ${COMPARE_ARGS} RESULT_VARIABLE IM_RESULT OUTPUT_VARIABLE IM_OUT)
+  if (IM_RESULT)
+    message(STATUS "ImageMagick failed to run")
+    message(STATUS "Using alternative image comparison")
     set(DIFFPNG 1)
+  elif (${IM_OUT} STREQUAL "0")
+     message(STATUS "magicktest1.png and magicktest2.png were incorrectly detected as identical")
+     message(STATUS "Using alternative image comparison")
+     set(DIFFPNG 1)
+  else()
+    message(STATUS "ImageMagick OK: Detected pixel difference of ${IM_OUT}")
   endif()
-  if (${DIFFPNG})
-    set(IMAGE_COMPARE_EXE ${CCBD}/diffpng)
-    set(COMPARATOR "--comparator=diffpng")
-    add_executable(diffpng diffpng.cpp ${CSD}/src/ext/lodepng/lodepng.cpp)
-    target_include_directories(diffpng PRIVATE ${CSD}/src/ext/lodepng)
-    message(STATUS "using diffpng for image comparison")
-  endif()
+endif()
+
+if (${DIFFPNG})
+  set(IMAGE_COMPARE_EXE ${CCBD}/diffpng)
+  set(COMPARATOR "--comparator=diffpng")
+  add_executable(diffpng diffpng.cpp ${CSD}/src/ext/lodepng/lodepng.cpp)
+  target_include_directories(diffpng PRIVATE ${CSD}/src/ext/lodepng)
+  message(STATUS "using diffpng for image comparison")
 endif()
 
 find_package(Lib3MF QUIET)
@@ -1385,8 +1334,6 @@ if(APPLE)
     # Z issues specific to edge shader
     openscad-viewoptions-edges_view-options-tests
     openscad-viewoptions-axes-scales-edges_view-options-tests
-    # issue3158 test failing on macos, see: https://github.com/openscad/openscad/issues/4509
-    opencsgtest_issue3158
     PROPERTIES DISABLED TRUE
   )
 elseif(UNIX)
